fix voice and fax ext

+++ src/main/java/godaddy/registry/jtoolkit2/se/ContactInfoResponse.java
@@ -90,9 +90,9 @@ public class ContactInfoResponse extends InfoResponse {
     private IntPostalInfo intPostalInfo;
     private LocalPostalInfo locPostalInfo;
     private String voice;
-    private int voiceX;
+    private long voiceX;
     private String fax;
-    private int faxX;
+    private long faxX;
     private String email;
     private String pw;
     private boolean discloseFlag;
@@ -100,8 +100,8 @@ public class ContactInfoResponse extends InfoResponse {
 
     public ContactInfoResponse() {
         super(StandardObjectType.CONTACT);
-        voiceX = -1;
-        faxX = -1;
+        voiceX = -1L;
+        faxX = -1L;
     }
 
     public String getID() {
@@ -120,7 +120,7 @@ public class ContactInfoResponse extends InfoResponse {
         return voice;
     }
 
-    public int getVoiceExtension() {
+    public long getVoiceExtension() {
         return voiceX;
     }
 
@@ -128,7 +128,7 @@ public class ContactInfoResponse extends InfoResponse {
         return fax;
     }
 
-    public int getFaxExtension() {
+    public long getFaxExtension() {
         return faxX;
     }
 
@@ -201,12 +201,12 @@ public class ContactInfoResponse extends InfoResponse {
             voice = xmlDoc.getNodeValue(CON_VOICE_EXPR);
             String voiceXStr = xmlDoc.getNodeValue(CON_VOICEX_EXPR);
             if (voiceXStr != null && voiceXStr.length() > 0) {
-                voiceX = Integer.parseInt(voiceXStr);
+                voiceX = Long.parseLong(voiceXStr);
             }
             fax = xmlDoc.getNodeValue(CON_FAX_EXPR);
             String faxXStr = xmlDoc.getNodeValue(CON_FAXX_EXPR);
             if (faxXStr != null && faxXStr.length() > 0) {
-                faxX = Integer.parseInt(faxXStr);
+                faxX = Long.parseLong(faxXStr);
             }
             email = xmlDoc.getNodeValue(CON_EMAIL_EXPR);
             pw = xmlDoc.getNodeValue(CON_PW_EXPR);
